multi agent task/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
faiss-cpu==1.8.0
PyMuPDF==1.23.8
sentence-transformers==2.2.2
groq==0.5.0
google-generativeai==0.3.2
arxiv==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.350
langchain-community==0.0.11
chromadb==0.4.22
pytest==7.4.3
httpx==0.25.2
reportlab==4.0.4
gradio==4.0.0